added IDLogger and modified build to include the external dependencies in the project

+++ log120/tp1/src/Line.java
@@ -6,13 +6,13 @@ public class Line extends Shape {
 	
   private Point p1, p2;
 	
-	public Line(long id, Color color, int x1, int y1, int x2, int y2) {
+	public Line(int id, Color color, int x1, int y1, int x2, int y2) {
 		super(id, color);
     this.p1 = new Point(x1, y1);
     this.p2 = new Point(x2, y2);
 	}
 
-  public Line(long id, Color color, Point p1, Point p2) {
+  public Line(int id, Color color, Point p1, Point p2) {
     super(id, color);
     this.p1 = p1;
     this.p2 = p2;
